@@ -1,26 +1,26 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" 
 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 <modelVersion>4.0.0</modelVersion>
-   <parent>
-	   <groupId>org.springframework.boot</groupId>
-	   <artifactId>spring-boot-starter-parent</artifactId>
-	   <version>2.7.5</version>
-	   <relativePath />
-   </parent>
-
+	<parent>
+		<groupId>org.springframework.boot</groupId>
+		<artifactId>spring-boot-starter-parent</artifactId>
+		<version>3.3.4</version>
+		<relativePath />
+	</parent>
 	<groupId>com.example</groupId>
 	<artifactId>where-for-dinner-api-gateway</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-
 	<name>${project.artifactId}</name>
 	<description>Where for Dinner services API gateway</description>
 	<inceptionYear>2022</inceptionYear>
-
 	<properties>
-        <java.version>11</java.version>
-		<spring-cloud.version>2021.0.5</spring-cloud.version>
-		<feign-reactor.version>3.1.1</feign-reactor.version>
-		<snakeyaml.version>1.32</snakeyaml.version>
+		<java.version>21</java.version>
+		<spring-cloud.version>2023.0.3</spring-cloud.version>
+		<spring-cloud-bindings.version>2.0.3</spring-cloud-bindings.version>
+		<okio-jvm.version>3.9.0</okio-jvm.version>
+		<cfenv.version>3.1.5</cfenv.version>	
+		<spring-cloud-services.version>4.1.2</spring-cloud-services.version>	
+		<xstream.version>1.4.20</xstream.version>				
 	</properties>
 	<dependencyManagement>
 		<dependencies>
