@@ -14,10 +14,13 @@
 	<name>where-for-dinner-availability</name>
 	<description>Where For Dinner service for storing and retriving establishment reservation openings.</description>
 	<properties>
-		<java.version>11</java.version>
-		<spring-cloud.version>2021.0.5</spring-cloud.version>
-		<spring-cloud-bindings.version>1.10.0</spring-cloud-bindings.version>
-		<snakeyaml.version>1.32</snakeyaml.version>
+		<java.version>21</java.version>
+		<spring-cloud.version>2023.0.3</spring-cloud.version>
+		<spring-cloud-bindings.version>2.0.3</spring-cloud-bindings.version>		
+		<springdoc.version>2.5.0</springdoc.version>	
+		<cfenv.version>3.1.5</cfenv.version>	
+		<spring-cloud-services.version>4.1.2</spring-cloud-services.version>
+		<xstream.version>1.4.20</xstream.version>				
 	</properties>
 	<dependencies>
 		<dependency>
