@@ -14,9 +14,12 @@
 	<name>where-for-dinner-notify</name>
 	<description>Where for Dinner notification application</description>
 	<properties>
-		<java.version>11</java.version>
-		<spring-cloud.version>2021.0.5</spring-cloud.version>
-		<snakeyaml.version>1.32</snakeyaml.version>
+		<java.version>21</java.version>		
+		<spring-cloud.version>2023.0.3</spring-cloud.version>
+		<spring-cloud-bindings.version>2.0.3</spring-cloud-bindings.version>
+		<cfenv.version>3.1.5</cfenv.version>	
+		<spring-cloud-services.version>4.1.2</spring-cloud-services.version>	
+		<xstream.version>1.4.20</xstream.version>	
 	</properties>
 	<dependencies>
 		<dependency>
