@@ -14,10 +14,12 @@
 	<name>where-for-dinner-search-proc</name>
 	<description>Where for Dinner search processor</description>
 	<properties>
-		<java.version>11</java.version>
-		<spring-cloud.version>2021.0.5</spring-cloud.version>
-		<feign-reactor.version>3.2.1</feign-reactor.version>
-		<snakeyaml.version>1.32</snakeyaml.version>
+		<java.version>21</java.version>
+		<spring-cloud.version>2023.0.3</spring-cloud.version>
+		<spring-cloud-bindings.version>2.0.3</spring-cloud-bindings.version>	
+		<cfenv.version>3.1.5</cfenv.version>	
+		<spring-cloud-services.version>4.1.2</spring-cloud-services.version>	
+		<xstream.version>1.4.20</xstream.version>			
 	</properties>
 	<dependencies>
 		<dependency>
