@@ -2,22 +2,7 @@ package com.java.example.tanzu.wherefordinner.model;
 
 import java.util.List;
 
-import lombok.Data;
-
-@Data
-public class SearchCriteria 
+public record SearchCriteria (String name, long startTime, long endTime, List<String> diningTypes, List<String> diningNames, String postalCode, int radius)
 {
-	private String name;	
-	
-	private long startTime;
-	
-	private long endTime;
-	
-	private List<String> diningTypes;
-	
-	private List<String> diningNames;
-	
-	private String postalCode;
-	
-	private int radius;
+
 }
