@@ -1,44 +1,9 @@
 package com.java.example.tanzu.wherefordinner.model;
 
-import java.util.ArrayList;
 import java.util.List;
 
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-import lombok.Data;
 
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Availability 
+public record Availability(String searchName, String diningName, String address, String locality, String region, String postalCode, String phoneNumber, 
+		String reservationURL, String requestSubject, String sendResultsTo, List<AvailabilityWindow> availabilityWindows)
 {	
-	private String searchName;
-
-	private String diningName;
-
-	private String address;
-	
-	private String locality;
-	
-	private String region;
-	
-	private String postalCode;
-	
-	private String phoneNumber;
-	
-	private String reservationURL;
-
-	private String requestSubject;
-	
-	private List<AvailabilityWindow> availabilityWindows = new ArrayList<>();
-	
-	@Data
-	@AllArgsConstructor
-	@NoArgsConstructor
-	public static class AvailabilityWindow
-	{	
-		private long startTime;
-		
-		private long endTime;
-	}
 }
