@@ -4,8 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
 
-import com.java.example.tanzu.wherefordinner.feign.CrawlerClient;
+import com.java.example.tanzu.wherefordinner.exchange.CrawlerClient;
 import com.java.example.tanzu.wherefordinner.model.Availability;
 import com.java.example.tanzu.wherefordinner.model.SearchCriteria;
 import com.java.example.tanzu.wherefordinner.searcher.Searcher;
@@ -26,16 +27,13 @@ public class CrawlerSearcher implements Searcher
 	@Override
 	public Flux<Availability> search(SearchCriteria crit) 
 	{
-		log.info("Making crawler dining search for dining search {}", crit.getName());
+		log.info("Making crawler dining search for dining search {}", crit.name());
 		
-		return crawlClient.search(crit.getDiningNames(), crit.getDiningTypes(), crit.getStartTime(), crit.getEndTime())
-			.map(avail -> 
-			{
-				avail.setSearchName(crit.getName());
-				avail.setRequestSubject(crit.getRequestSubject());
-				
-				return avail;
-			});
+		return crawlClient.search(StringUtils.hasText(crit.diningNames()) ? crit.diningNames() : "", 
+				StringUtils.hasText(crit.diningTypes()) ? crit.diningTypes() : "", 
+				crit.startTime(), crit.endTime())
+			.map(avail -> new Availability(crit.name(), avail.diningName(), avail.address(), avail.locality(), avail.region(), avail.postalCode(), avail.phoneNumber(),
+					avail.reservationURL(), crit.requestSubject(), crit.sendResultsTo(), avail.availabilityWindows()));
 			
 	}
 
