@@ -2,7 +2,6 @@ package com.java.example.tanzu.wherefordinner.searcher.impl;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.Random;
 import java.util.stream.Collectors;
@@ -14,12 +13,15 @@ import org.springframework.util.StringUtils;
 
 import com.java.example.tanzu.wherefordinner.config.StaticDiningAvailability;
 import com.java.example.tanzu.wherefordinner.model.Availability;
+import com.java.example.tanzu.wherefordinner.model.AvailabilityWindow;
 import com.java.example.tanzu.wherefordinner.model.SearchCriteria;
 import com.java.example.tanzu.wherefordinner.searcher.Searcher;
 
+import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Flux;
 
 @Component
+@Slf4j
 public class LocalRandomSearcher implements Searcher
 {
 
@@ -35,19 +37,21 @@ public class LocalRandomSearcher implements Searcher
 	{		
 		// if there is a list of dining names, we want to pick from that list
 		final var diningFlux = createRandomDinings(crit);
-			
+		
 		return diningFlux.map(dining -> createRandomAvailability(dining, crit))	;		
 	}
 
 	protected Flux<StaticDiningAvailability.Establishment> createRandomDinings(SearchCriteria crit)
 	{
+		log.debug("Creating random dining selections");
+		
 		List<StaticDiningAvailability.Establishment> bucket = null;
 		var rn = new Random();
 		var numDinings = 0;
 		
-		if (StringUtils.hasText(crit.getDiningNames()))
+		if (StringUtils.hasText(crit.diningNames()))
 		{
-			var dinNames = Arrays.asList(crit.getDiningNames().trim().split(","));
+			var dinNames = Arrays.asList(crit.diningNames().trim().split(","));
 			
 			var filteredDining = staticDining.getEstablishments().stream()
 			.filter(estab -> isEstablishmentInList(estab.getDiningName(), dinNames))
@@ -86,12 +90,14 @@ public class LocalRandomSearcher implements Searcher
 	
 	protected Availability createRandomAvailability(StaticDiningAvailability.Establishment dining, SearchCriteria crit)
 	{
+		log.debug("Creating random dining availability");
+		
 		var rn = new Random();
 		
 		// the chance that no reservations are available will be 20%
 		final var isTimesAvailable = (rn.nextInt(10) > 1);
 		
-		final var avail = new Availability();
+		final List<AvailabilityWindow> availWindows = new ArrayList<>();
 		
 		if (isTimesAvailable)
 		{
@@ -99,25 +105,25 @@ public class LocalRandomSearcher implements Searcher
 			final var times = new long[2];
 			
 			// select a random start time... the start time should not be less than an hour before the end time
-			var windowSize = crit.getEndTime() - crit.getStartTime();
+			var windowSize = crit.endTime() - crit.startTime();
 			
 			// if the start and stop times are less than a 1 hour window, then just select
 			// the start and end times as the window
 			if (windowSize <= HOUR)
 			{
-				times[0] =  crit.getStartTime();
-				times[1] = crit.getEndTime();
+				times[0] =  crit.startTime();
+				times[1] = crit.endTime();
 			}
 			else
 			{
 				// find a random start time
-				var max = crit.getEndTime() - HOUR;
-				var min = crit.getStartTime();
+				var max = crit.endTime() - HOUR;
+				var min = crit.startTime();
 				var startTime = min + (long) (Math.random() * (max - min));
 				startTime = roundToHalfHour(startTime - HALF_HOUR);
 				
 				// find a random end time time
-				max = crit.getEndTime();
+				max = crit.endTime();
 				min = startTime + HALF_HOUR;
 				var endTime = min + (long) (Math.random() * (max - min));
 				endTime = roundToHalfHour(endTime - HALF_HOUR);
@@ -126,20 +132,13 @@ public class LocalRandomSearcher implements Searcher
 				times[1] = endTime;	
 			}
 			
-			avail.setAvailabilityWindows(Collections.singletonList(new Availability.AvailabilityWindow(times[0], times[1])));
+			availWindows.add(new AvailabilityWindow(times[0], times[1]));
 		}
 		
-		avail.setSearchName(crit.getName());
-		avail.setDiningName(dining.getDiningName());
-		avail.setAddress(dining.getAddress());
-		avail.setLocality(dining.getLocality());
-		avail.setPhoneNumber(dining.getPhoneNumber());
-		avail.setPostalCode(dining.getPostalCode());
-		avail.setRegion(dining.getRegion());
-		avail.setReservationURL(dining.getReservationURL());
-		avail.setRequestSubject(crit.getRequestSubject());
-		
-		return avail;
+		return new Availability(crit.name(), dining.getDiningName(), dining.getAddress(), dining.getLocality(), dining.getRegion(), dining.getPostalCode(), dining.getPhoneNumber(),
+				dining.getReservationURL(), crit.requestSubject(), crit.sendResultsTo(), availWindows);
+
+
 	}
 	
 	protected long roundToHalfHour(long time)
