@@ -41,8 +41,8 @@ function App() {
               </div>      	
             </div>
             <div className="container welcomeContent">
-                      The Where For Dinner Dining Availability application continously searches for dining availability based on desired search
-                      parameters.  Once a search is submitted, the Where For Dinner system will continously scour different sources for dining
+                      The Where For Dinner Dining Availability application continuously searches for dining availability based on desired search
+                      parameters.  Once a search is submitted, the Where For Dinner system will continuously scour different sources for dining
                       availability that match your preferences and display dining availability in near real time.  Searches will continue even 
                       after you leave the application, so you may return a later time to check to see if any new availability has been found.
             </div>          
@@ -53,8 +53,8 @@ function App() {
                 </div>
               </Route>
               <Route path="/diningsearch">
-                <form action="/scg-logout" method="POST" id="form">
-                  <input type="hidden" id="var1" name="var1" value=""/>
+                <form action="/scg-logout" method="GET" id="form">
+                  <input type="hidden" id="logout" name="redirect" value="/home"/>
                   {secEnabled ?                  
                     <div align="center" id="logout">
                       <button>Logout</button>
